+++ pkg/diag/errors.go
@@ -67,3 +67,16 @@ func GetResourceToRefreshCouldNotBeFoundDidYouForgetError() *Diag {
 	return newError("", 2011, "Resource to refresh '%v' could not be found in the stack. "+
 		"Did you forget to escape $ in your shell?")
 }
+
+func GetResourceToDeleteCouldNotBeFoundError() *Diag {
+	return newError("", 2012, "Resource to delete '%v' could not be found in the stack.")
+}
+
+func GetResourceToDeleteCouldNotBeFoundDidYouForgetError() *Diag {
+	return newError("", 2013, "Resource to delete '%v' could not be found in the stack. "+
+		"Did you forget to escape $ in your shell?")
+}
+
+func GetCannotDeleteParentResourceWithoutAlsoDeletingChildError() *Diag {
+	return newError("", 2014, "Cannot delete parent resource '%v' without also deleting child '%v'.")
+}