+++ pkg/engine/lifecycle_test.go
@@ -441,6 +441,16 @@ func (p *TestPlan) Run(t *testing.T, snapshot *deploy.Snapshot) *deploy.Snapshot
 			continue
 		}
 
+		if res != nil {
+			if res.IsBail() {
+				t.Logf("Got unexpected bail result")
+				t.FailNow()
+			} else {
+				t.Logf("Got unexpected error result: %v", res.Error())
+				t.FailNow()
+			}
+		}
+
 		assert.Nil(t, res)
 	}
 
@@ -1556,17 +1566,18 @@ func TestRefreshWithDelete(t *testing.T) {
 	}
 }
 
-func pickURN(urns []resource.URN, target string) resource.URN {
-	switch target {
-	case "resA":
-		return urns[0]
-	case "resB":
-		return urns[1]
-	case "resC":
-		return urns[2]
-	default:
-		panic("Invalid target: " + target)
+func pickURN(t *testing.T, urns []resource.URN, names []string, target string) resource.URN {
+	assert.Equal(t, len(urns), len(names))
+	assert.Contains(t, names, target)
+
+	for i, name := range names {
+		if name == target {
+			return urns[i]
+		}
 	}
+
+	t.Fatalf("Could not find target: %v in %v", target, names)
+	return ""
 }
 
 // Tests that dependencies are correctly rewritten when refresh removes deleted resources.
@@ -1600,7 +1611,7 @@ func validateRefreshDeleteCombination(t *testing.T, names []string, targets []st
 
 	t.Logf("Refreshing targets: %v", targets)
 	for _, target := range targets {
-		refreshTargets = append(p.Options.RefreshTargets, pickURN(urns, target))
+		refreshTargets = append(refreshTargets, pickURN(t, urns, names, target))
 	}
 
 	p.Options.RefreshTargets = refreshTargets
@@ -1757,7 +1768,7 @@ func validateRefreshBasicsCombination(t *testing.T, names []string, targets []st
 	refreshTargets := []resource.URN{}
 
 	for _, target := range targets {
-		refreshTargets = append(p.Options.RefreshTargets, pickURN(urns, target))
+		refreshTargets = append(p.Options.RefreshTargets, pickURN(t, urns, names, target))
 	}
 
 	p.Options.RefreshTargets = refreshTargets
@@ -4333,4 +4344,209 @@ func TestImportUpdatedID(t *testing.T) {
 			t.Fatalf("unexpected resource %v", urn)
 		}
 	}
+}
+
+func TestDeleteTarget(t *testing.T) {
+	names := []string{"A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L"}
+
+	// Try refreshing a stack with combinations of the above resources as target to destroy.
+	subsets := combinations.All(names)
+
+	for _, subset := range subsets {
+		// limit to up to 3 resources to destroy.  This keeps the test running time under
+		// control as it only generates a few hundred combinations instead of several thousand.
+		if len(subset) <= 3 {
+			deleteSpecificTargets(t, names, subset, func(urns []resource.URN, deleted map[resource.URN]bool) {})
+		}
+	}
+
+	deleteSpecificTargets(t, names, []string{"A"}, func(urns []resource.URN, deleted map[resource.URN]bool) {
+		// when deleting 'A' we expect A, B, C, E, F, and K to be deleted
+		assert.Equal(t, map[resource.URN]bool{
+			pickURN(t, urns, names, "A"): true,
+			pickURN(t, urns, names, "B"): true,
+			pickURN(t, urns, names, "C"): true,
+			pickURN(t, urns, names, "E"): true,
+			pickURN(t, urns, names, "F"): true,
+			pickURN(t, urns, names, "K"): true,
+		}, deleted)
+	})
+}
+
+func deleteSpecificTargets(
+	t *testing.T, names []string, targets []string,
+	validate func(urns []resource.URN, deleted map[resource.URN]bool)) {
+
+	//             A
+	//    _________|_________
+	//    B        C        D
+	//          ___|___  ___|___
+	//          E  F  G  H  I  J
+	//             |__|
+	//             K  L
+
+	p := &TestPlan{}
+
+	const resType = "pkgA:m:typA"
+	type propertyDependencies map[resource.PropertyKey][]resource.URN
+
+	urnA := p.NewProviderURN("pkgA", names[0], "")
+	urnB := p.NewURN(resType, names[1], "")
+	urnC := p.NewProviderURN("pkgA", names[2], "")
+	urnD := p.NewProviderURN("pkgA", names[3], "")
+	urnE := p.NewURN(resType, names[4], "")
+	urnF := p.NewURN(resType, names[5], "")
+	urnG := p.NewURN(resType, names[6], "")
+	urnH := p.NewURN(resType, names[7], "")
+	urnI := p.NewURN(resType, names[8], "")
+	urnJ := p.NewURN(resType, names[9], "")
+	urnK := p.NewURN(resType, names[10], "")
+	urnL := p.NewURN(resType, names[11], "")
+	urns := []resource.URN{
+		urnA, urnB, urnC, urnD, urnE, urnF,
+		urnG, urnH, urnI, urnJ, urnK, urnL,
+	}
+
+	newResource := func(urn resource.URN, id resource.ID, provider string, dependencies []resource.URN,
+		propertyDeps propertyDependencies, outputs resource.PropertyMap) *resource.State {
+
+		inputs := resource.PropertyMap{}
+		for k := range propertyDeps {
+			inputs[k] = resource.NewStringProperty("foo")
+		}
+
+		return &resource.State{
+			Type:                 urn.Type(),
+			URN:                  urn,
+			Custom:               true,
+			Delete:               false,
+			ID:                   id,
+			Inputs:               inputs,
+			Outputs:              outputs,
+			Dependencies:         dependencies,
+			Provider:             provider,
+			PropertyDependencies: propertyDeps,
+		}
+	}
+
+	old := &deploy.Snapshot{
+		Resources: []*resource.State{
+			newResource(urnA, "0", "", nil, nil, resource.PropertyMap{"A": resource.NewStringProperty("foo")}),
+			newResource(urnB, "1", string(urnA)+"::0", nil, nil, nil),
+			newResource(urnC, "2", "",
+				[]resource.URN{urnA},
+				propertyDependencies{"A": []resource.URN{urnA}},
+				resource.PropertyMap{"A": resource.NewStringProperty("bar")}),
+			newResource(urnD, "3", "",
+				[]resource.URN{urnA},
+				propertyDependencies{"B": []resource.URN{urnA}}, nil),
+			newResource(urnE, "4", string(urnC)+"::2", nil, nil, nil),
+			newResource(urnF, "5", "",
+				[]resource.URN{urnC},
+				propertyDependencies{"A": []resource.URN{urnC}}, nil),
+			newResource(urnG, "6", "",
+				[]resource.URN{urnC},
+				propertyDependencies{"B": []resource.URN{urnC}}, nil),
+			newResource(urnH, "4", string(urnD)+"::3", nil, nil, nil),
+			newResource(urnI, "5", "",
+				[]resource.URN{urnD},
+				propertyDependencies{"A": []resource.URN{urnD}}, nil),
+			newResource(urnJ, "6", "",
+				[]resource.URN{urnD},
+				propertyDependencies{"B": []resource.URN{urnD}}, nil),
+			newResource(urnK, "7", "",
+				[]resource.URN{urnF, urnG},
+				propertyDependencies{"A": []resource.URN{urnF, urnG}}, nil),
+			newResource(urnL, "8", "",
+				[]resource.URN{urnF, urnG},
+				propertyDependencies{"B": []resource.URN{urnF, urnG}}, nil),
+		},
+	}
+
+	loaders := []*deploytest.ProviderLoader{
+		deploytest.NewProviderLoader("pkgA", semver.MustParse("1.0.0"), func() (plugin.Provider, error) {
+			return &deploytest.Provider{
+				DiffConfigF: func(urn resource.URN, olds, news resource.PropertyMap,
+					ignoreChanges []string) (plugin.DiffResult, error) {
+					if !olds["A"].DeepEquals(news["A"]) {
+						return plugin.DiffResult{
+							ReplaceKeys:         []resource.PropertyKey{"A"},
+							DeleteBeforeReplace: true,
+						}, nil
+					}
+					return plugin.DiffResult{}, nil
+				},
+				DiffF: func(urn resource.URN, id resource.ID,
+					olds, news resource.PropertyMap, ignoreChanges []string) (plugin.DiffResult, error) {
+
+					if !olds["A"].DeepEquals(news["A"]) {
+						return plugin.DiffResult{ReplaceKeys: []resource.PropertyKey{"A"}}, nil
+					}
+					return plugin.DiffResult{}, nil
+				},
+			}, nil
+		}),
+	}
+
+	program := deploytest.NewLanguageRuntime(func(_ plugin.RunInfo, monitor *deploytest.ResourceMonitor) error {
+		register := func(urn resource.URN, provider string, inputs resource.PropertyMap) resource.ID {
+			_, id, _, err := monitor.RegisterResource(urn.Type(), string(urn.Name()), true, deploytest.ResourceOptions{
+				Provider: provider,
+				Inputs:   inputs,
+			})
+			assert.NoError(t, err)
+			return id
+		}
+
+		idA := register(urnA, "", resource.PropertyMap{"A": resource.NewStringProperty("bar")})
+		register(urnB, string(urnA)+"::"+string(idA), nil)
+		idC := register(urnC, "", nil)
+		idD := register(urnD, "", nil)
+		register(urnE, string(urnC)+"::"+string(idC), nil)
+		register(urnF, "", nil)
+		register(urnG, "", nil)
+		register(urnH, string(urnD)+"::"+string(idD), nil)
+		register(urnI, "", nil)
+		register(urnJ, "", nil)
+		register(urnK, "", nil)
+		register(urnL, "", nil)
+
+		return nil
+	})
+
+	p.Options.host = deploytest.NewPluginHost(nil, nil, program, loaders...)
+
+	destroyTargets := []resource.URN{}
+	for _, target := range targets {
+		destroyTargets = append(destroyTargets, pickURN(t, urns, names, target))
+	}
+
+	p.Options.DestroyTargets = destroyTargets
+	t.Logf("Destroying targets: %v", destroyTargets)
+
+	p.Steps = []TestStep{{
+		Op:            Destroy,
+		ExpectFailure: false,
+		Validate: func(project workspace.Project, target deploy.Target, j *Journal,
+			evts []Event, res result.Result) result.Result {
+
+			assert.Nil(t, res)
+			assert.True(t, len(j.Entries) > 0)
+
+			deleted := make(map[resource.URN]bool)
+			for _, entry := range j.Entries {
+				assert.Equal(t, deploy.OpDelete, entry.Step.Op())
+				deleted[entry.Step.URN()] = true
+			}
+
+			for _, target := range p.Options.DestroyTargets {
+				assert.Contains(t, deleted, target)
+			}
+
+			validate(urns, deleted)
+			return res
+		},
+	}}
+
+	p.Run(t, old)
 }